Added more 64-bit support

+++ utils/RCBot2_meta/bot_css_mod.cpp
@@ -318,9 +318,9 @@ edict_t *CCounterStrikeSourceMod::getRandomHostage()
         }
     }
 
-    if(!temp.empty())
-    {   
-        return INDEXENT(temp.at(randomInt(0, static_cast<int>(temp.size() - 1))).GetEntryIndex());
+    if (!temp.empty())
+    {
+        return INDEXENT(temp.at(static_cast<std::size_t>(randomInt(0, static_cast<int>(temp.size() - 1)))).GetEntryIndex());
     }
 
     return nullptr;
